app/process_microservice/requirements.txt
pytesseract
Pillow
fastapi
pydantic
pydantic-settings
uvicorn
beanie
requests
pymongo
pytest 
pytest_asyncio
httpx